chapterDT/requirements.txt
alabaster==0.7.12
alembic==1.2.1
altgraph==0.16.1
asn1crypto==1.2.0
astroid==2.3.2
async-generator==1.10
attrs==19.3.0
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
buildozer==0.40.dev0
certifi==2019.9.11
certipy==0.1.3
cffi==1.13.0
chardet==3.0.4
Click==7.0
cmake==3.15.3
colorama==0.4.1
cryptography==2.8
cx-Oracle==7.2.3
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.6
dlib==19.18.0
docutils==0.15.2
entrypoints==0.3
facebook-sdk==3.1.0
Flask==1.1.1
future==0.18.1
graphviz==0.13
hide-code==0.5.2
html5lib==1.0.1
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
imagesize==1.1.0
importlib-metadata==0.23
IP2Location==8.4.1
ipykernel==5.1.2
ipyparallel==6.2.4
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.0
jupyterhub==1.0.0
Kivy==1.11.1
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.17
kivy-deps.sdl2==0.1.22
Kivy-Garden==0.1.4
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lxml==4.4.1
macholib==1.11
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
moviepy==1.0.1
mpmath==1.1.0
mysql-connector==2.2.9
nbconvert==5.6.0
nbformat==4.4.0
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.1.1.26
packaging==19.2
pafy==0.5.4
pamela==1.0.0
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
pdfkit==0.6.1
pefile==2019.4.18
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pip-review==1.0
pipenv==2018.11.26
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psutil==5.6.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pydot==1.4.1
pydotplus==2.0.2
pyenv-win==1.2.2
pygame==1.9.6
Pygments==2.4.2
PyInstaller==3.5
pylint==2.4.3
pymssql==2.1.4
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
Pypubsub==4.0.3
PyQt5==5.13.1
PyQt5-sip==12.7.0
pyrsistent==0.15.4
python-dateutil==2.8.0
python-docx==0.8.10
python-editor==1.0.4
python-oauth2==1.1.1
python-vlc==3.0.7110
pytz==2019.3
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
rope==0.14.0
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
sh==1.12.14
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==2.0.0
spark-parser==1.8.9
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.10
sqlparse==0.3.0
style==1.1.6
sympy==1.4
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.0
uncompyle6==3.5.0
update==0.0.1
urllib3==1.25.6
virtualenv==16.7.6
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
wxPython==4.0.6
xdis==4.1.0
youtube-dl==2019.10.16
zipp==0.6.0
